+++ src/main/fi/testbed2/app/MainApplication.java
@@ -10,7 +10,7 @@ import fi.testbed2.data.TestbedParsedPage;
 
 public class MainApplication extends Application {
 
-	public static final String LOG_IDENTIFIER = "TestbedViewer2";
+    public static final String LOG_IDENTIFIER = "TestbedViewer2";
 
 	public static final int RESULT_ERROR = Activity.RESULT_FIRST_USER;
     public static final int RESULT_OK = Activity.RESULT_OK;
@@ -23,20 +23,28 @@ public class MainApplication extends Application {
     public static final String PREF_MAP_NUMBER_OF_IMAGES = "PREF_MAP_NUMBER_OF_IMAGES";
     public static final String PREF_ORIENTATION_PREFERENCE_KEY_PREFIX = "PREFERENCE_ANIM_BOUNDS_ORIENTATION_";
 
-    private static TestbedParsedPage testbedParsedPage;
     private static LruCache<String, Bitmap> imageCache;
+    private static LruCache<String, TestbedParsedPage> pageCache;
+
+    private static String PAGE_CACHE_KEY = "TESTBED_PAGE";
 
     public static TestbedParsedPage getTestbedParsedPage() {
-        return testbedParsedPage;
+        return pageCache.get(PAGE_CACHE_KEY);
     }
 
     public static void setTestbedParsedPage(TestbedParsedPage testbedParsedPage) {
-        MainApplication.testbedParsedPage = testbedParsedPage;
+        if (testbedParsedPage==null) {
+            pageCache.remove(PAGE_CACHE_KEY);
+            System.gc();
+        } else {
+            pageCache.put(PAGE_CACHE_KEY, testbedParsedPage);
+        }
     }
 
     @Override
     public void onCreate() {
         initImageCache();
+        initPageCache();
     }
 
     private void initImageCache() {
@@ -65,14 +73,34 @@ public class MainApplication extends Application {
 
     }
 
+    private void initPageCache() {
+
+        final int memClass = ((ActivityManager) this.getSystemService(
+                Context.ACTIVITY_SERVICE)).getMemoryClass();
+
+        final int cacheSizeInBytes = 1024 * 1024 * memClass / 8;
+
+        pageCache = new LruCache<String, TestbedParsedPage>(cacheSizeInBytes);
+
+    }
+
     public static void clearData() {
-        testbedParsedPage = null;
+        clearPageCache();
         clearImageCache();
+        System.gc();
+    }
+
+    public static void deleteBitmapCacheEntry(String key) {
+        imageCache.remove(key);
         System.gc();
     }
 
     private static void clearImageCache() {
         imageCache.evictAll();
+    }
+
+    private static void clearPageCache() {
+        pageCache.evictAll();
     }
 
     public static void addBitmapToImageCache(String key, Bitmap bitmap) {